mcp-memory-local/services/shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-memory-shared"
version = "0.1.0"
description = "Shared library for MCP memory services"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
